[JENKINS-49144] JobInvocation: stop serializing locks
Locks can't be serialized since JEP-200 (serialization whitelist instead
of a blacklist) and it makes no sense to do so anyway. Also, I think
it's not necessary to be able to call buildFinalized/waitForFinalization
on deserialized instances of JobInvocation as these only make sense
during the flow run itself, not when viewing results.

JIRA: SETI-1893
JIRA: SETI-1273